backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
psycopg2-binary==2.9.9
pretty_midi==0.2.10
mido==1.3.2
music21==9.1.0
numpy==1.26.4
gunicorn==22.0.0
